refactor(front): move data_source DTO types out of lib/api into types/api
Fourth batch of the lib/api DTO migration: internal data_source endpoints.
Transport request/response DTO types and zod schemas move to the mirrored path
under types/api/*, leaving business-logic functions and domain types in place.

- 4 pure-contract files moved wholesale: data_source_view_tags,
  data_sources/bot_data_sources, data_sources/managed_config,
  data_sources/managed_notion.
- 3 files trimmed (DTOs extracted, logic/domain stay): agent_data_sources
  (keeps DataSourcesUsageByAgent), datasets (drops unused zod import),
  data_sources/managed_permissions (keeps ManagedPermissionsResponse/Error
  domain types; back-imports ManagedPermissionsQuery used as a fn param).
- 15 consumers updated across front and front-api; 5 mixed-import files split.

data_sources.ts and data_source_view.ts are intentionally excluded: they have
public v1 consumers and belong in the SDK, handled in a later pass.
