+++ front-api/routes/w/[wId]/data_sources/[dsId]/usage.ts
@@ -1,6 +1,6 @@
-import type { GetDataSourceUsageResponseBody } from "@app/lib/api/agent_data_sources";
 import { getDataSourceUsage } from "@app/lib/api/agent_data_sources";
 import { DataSourceResource } from "@app/lib/resources/data_source_resource";
+import type { GetDataSourceUsageResponseBody } from "@app/types/api/agent_data_sources";
 import { workspaceApp } from "@front-api/middlewares/ctx";
 import type { HandlerResult } from "@front-api/middlewares/utils";
 import { apiError } from "@front-api/middlewares/utils";
