+++ front-api/routes/w/[wId]/spaces/[spaceId]/apps/[aId]/datasets/index.ts
@@ -1,16 +1,14 @@
 import config from "@app/lib/api/config";
-import type {
-  GetDatasetsResponseBody,
-  PostDatasetResponseBody,
-} from "@app/lib/api/datasets";
-import {
-  getDatasets,
-  PostDatasetRequestBodySchema,
-} from "@app/lib/api/datasets";
+import { getDatasets } from "@app/lib/api/datasets";
 import { checkDatasetData } from "@app/lib/datasets";
 import { AppResource } from "@app/lib/resources/app_resource";
 import { DatasetModel } from "@app/lib/resources/storage/models/apps";
 import logger from "@app/logger/logger";
+import type {
+  GetDatasetsResponseBody,
+  PostDatasetResponseBody,
+} from "@app/types/api/datasets";
+import { PostDatasetRequestBodySchema } from "@app/types/api/datasets";
 import { CoreAPI } from "@app/types/core/core_api";
 import type { APIErrorResponse } from "@app/types/error";
 import { isString } from "@app/types/shared/utils/general";
