+++ front/lib/api/datasets.ts
@@ -6,32 +6,6 @@ import logger from "@app/logger/logger";
 import type { AppType } from "@app/types/app";
 import { CoreAPI } from "@app/types/core/core_api";
 import type { DatasetSchema, DatasetType } from "@app/types/dataset";
-import { z } from "zod";
-
-export type GetDatasetsResponseBody = {
-  datasets: DatasetType[];
-};
-
-export type PostDatasetResponseBody = {
-  dataset: DatasetType;
-};
-
-export type GetDatasetResponseBody = { dataset: DatasetType };
-
-export const PostDatasetRequestBodySchema = z.object({
-  dataset: z.object({
-    name: z.string(),
-    description: z.string().nullable(),
-    data: z.array(z.record(z.string(), z.any())),
-  }),
-  schema: z.array(
-    z.object({
-      key: z.string(),
-      type: z.enum(["string", "number", "boolean", "json"]),
-      description: z.string().nullable(),
-    })
-  ),
-});
 
 export async function getDatasets(
   auth: Authenticator,
