+++ front/types/api/agent_data_sources.ts
@@ -1,0 +1,5 @@
+import type { AgentsUsageType } from "@app/types/data_source";
+
+export type GetDataSourceUsageResponseBody = {
+  usage: AgentsUsageType;
+};
