+++ front/types/api/data_sources/managed_config.ts
